+++ test/end-to-end/offer.js
@@ -57,7 +57,6 @@ describe("Offer endpoint tests", () => {
         password: "password123",
     };
 
-
     beforeAll(async () => {
         await Company.deleteMany({});
         test_company = await Company.create({
@@ -3038,146 +3037,6 @@ describe("Offer endpoint tests", () => {
                         expect(res2.body.results[0].title).toBe(test_offer.title);
                     });
                 });
-            });
-        });
-    });
-
-    describe("GET /offers/company/:companyId", () => {
-        beforeAll(async () => {
-            await Offer.deleteMany({});
-        });
-
-        describe("Id Validation", () => {
-            test("should fail if requested an invalid companyId", async () => {
-                const companyId = "123";
-                const res = await request()
-                    .get(`/offers/company/${companyId}`);
-
-                expect(res.body.errors[0]).toHaveProperty("msg", ValidationReasons.OBJECT_ID);
-            });
-
-            test("should fail if there isn't a company with that id", async () => {
-                const missingCompanyId = "60ddb0bb2849830020883f91";
-                const res = await request().get(`/offers/company/${missingCompanyId}`);
-
-                expect(res.body.errors[0]).toHaveProperty("msg", ValidationReasons.COMPANY_NOT_FOUND(missingCompanyId));
-            });
-        });
-
-        describe("Get offer by companyId", () => {
-            const test_offers = [{}, {}, {}, {}];
-            const test_agent = agent();
-
-            beforeAll(async () => {
-                await Offer.deleteMany({});
-
-                const createOffer = async (offer) => {
-                    const { _id, owner, ownerName, ownerLogo } = await Offer.create({
-                        ...offer,
-                        owner: test_company._id.toString(),
-                        ownerName: test_company.name,
-                        ownerLogo: test_company.logo,
-                    });
-                    return {
-                        ...offer,
-                        owner: owner.toString(),
-                        ownerName,
-                        ownerLogo,
-                        _id: _id.toString()
-                    };
-                };
-
-                (await Promise.all(test_offers
-                    .map((_, i) => createOffer({ ...generateTestOffer(), isHidden: i === 2 }))))
-                    .forEach((elem, i) => {
-                        test_offers[i] = elem;
-                    });
-            });
-
-            test("should return hidden company offers as company", async () => {
-                // Login wiht test_user_company
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_company)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent.get(`/offers/company/${test_company._id}`);
-                expect(res.status).toBe(HTTPStatus.OK);
-
-                const extractedData = res.body;
-                expect(extractedData.map((offer) => offer._id).sort())
-                    .toMatchObject(
-                        test_offers.map((offer) => offer._id).sort()
-                    );
-
-                // Logout
-                await test_agent
-                    .del("/auth/login")
-                    .expect(HTTPStatus.OK);
-            });
-
-            test("should return non-hidden offers", async () => {
-                const res = await test_agent.get(`/offers/company/${test_company._id}`);
-                expect(res.status).toBe(HTTPStatus.OK);
-
-                const extractedData = res.body;
-                expect(extractedData.map((offer) => offer._id).sort())
-                    .toMatchObject(
-                        test_offers.filter((offer) => offer.isHidden === false).map((offer) => offer._id).sort()
-                    );
-            });
-
-            test("should return non-hidden offers, even if target owner is set", async () => {
-                const res = await test_agent
-                    .get(`/offers/company/${test_company._id}`)
-                    .send({
-                        owner: test_company._id
-                    });
-
-                expect(res.status).toBe(HTTPStatus.OK);
-
-                const extractedData = res.body;
-                expect(extractedData.map((offer) => offer._id).sort())
-                    .toMatchObject(
-                        test_offers.filter((offer) => offer.isHidden === false).map((offer) => offer._id).sort()
-                    );
-            });
-
-            test("should return hidden company offers as admin", async () => {
-                // Login with test_user_company
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_admin)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent.get(`/offers/company/${test_company._id}`);
-                expect(res.status).toBe(HTTPStatus.OK);
-
-                const extractedData = res.body;
-                expect(extractedData.map((offer) => offer._id).sort())
-                    .toMatchObject(
-                        test_offers.map((offer) => offer._id).sort()
-                    );
-
-                // Logout
-                await test_agent
-                    .del("/auth/login")
-                    .expect(HTTPStatus.OK);
-            });
-
-            test("should return hidden company offers with god token", async () => {
-                // Send request with god token
-                const res = await test_agent
-                    .get(`/offers/company/${test_company._id}`)
-                    .send(withGodToken());
-
-                expect(res.status).toBe(HTTPStatus.OK);
-
-                const extractedData = res.body;
-                expect(extractedData.map((offer) => offer._id).sort())
-                    .toMatchObject(
-                        test_offers.map((offer) => offer._id).sort()
-                    );
             });
         });
     });
